Added new Android Spawner block for spawning rogue androids in teams. Added new weapon Ion Sniper. Added new weapon module sniper scope. Implemented custom weapon model metadata for module attachment positions. Implemented proper per tick weapon fire synchronisation. Now phaser bolts get compensated for network lag. Fixed mouse position on GUI element groups. Added zoom modification to weapons. Separated weapon color modules and scopes to own interfaces. Rogue android no can have paths they can follow. Used mainly by Android spawners for PVE minigames. Improved Rogue android AI. Ranged Rogue androids now run away from close compat. Phaser Bolts now hurt only entities not on the same team. Fixed rogue androids not spawning with random armor. Rogue androids now avoid water and can't swim. IWeaponModule now modifies weapon stats directly This allows for advanced stat modification. Plasma Bolt Knockback can now be controlled. Did some rebalancing on energy weapon damage and speed.

+++ src/main/java/matteroverdrive/api/weapon/WeaponShot.java
@@ -39,7 +39,7 @@ public class WeaponShot
 
     public WeaponShot(ByteBuf buf)
     {
-        this(buf.readInt(), buf.readFloat(), buf.readFloat(), buf.readInt(),buf.readByte());
+        this(buf.readInt(), buf.readFloat(), buf.readFloat(), buf.readInt(),buf.readShort());
         setCount(buf.readByte());
     }
 
@@ -58,7 +58,7 @@ public class WeaponShot
         buf.writeFloat(damage);
         buf.writeFloat(accuracy);
         buf.writeInt(color);
-        buf.writeByte(range);
+        buf.writeShort(range);
         buf.writeByte(count);
     }
